requirements-screenshot.txt
accelerate
apex==0.9.10.dev0
beautifulsoup4==4.12.2
cssutils==2.9.0
datasets
google-ai-generativelanguage==0.4.0
google-api-core==2.16.0
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
html5validator==0.4.2
huggingface-hub
matplotlib==3.8.1
matplotlib-inline==0.1.6
multiprocess==0.70.16
numpy==1.26.1
greenlet==3.0.3
tqdm==4.66.2
pillow==10.2.0
pydantic==2.6.3
httpx==0.27.0
distro==1.9.0
pyparsing==3.1.2
cycler==0.12.1
threadpoolctl==3.3.0
ftfy==6.1.3
regex==2023.12.25
scipy
scikit-learn
colormath
playwright
opencv-python
nltk
seaborn
retry
pyee
